-improved variable rate tracing
basically, now variance is computed during a-trous filters. This has better performance because we can skip separate variance estimation passes (and in a-trous we have to compute variance anyways, as those filters are variance driven) and results in better quality. That's because variance from a trous is not just a mean variance from neighboorhood - it's also decreased with more samples averages (same as normal variance - 2x more samples give 2x smaller variance). This is nice because it makes variance estimation also kinda aware of geometry and materials complexity which results in better estimation