-implement boost to specular variance for very smooth surfaces -fix temporal reprojection during temporal resampling
generally this CL should improve bahavior for materials that have very low roughness and very low metallic. These look normally good, because diffuse reflections contribute usually way more than specular but this can break when looking at grazing angle. In theory, specular influence should be larger in those cases but often it's still not large enough to trace a lot of specular rays. This is a problem because with those surfaces any instability in specular reflections can be very visible. Such materials are a common case on Pica Pica scene